requirements.txt
flask
gunicorn
gpytranslate
googletrans
pytgcalls==3.0.0.dev24
openai
uvloop
ffmpeg-python
git+https://github.com/yt-dlp/yt-dlp.git
youtube-search
youtube-search-python
APScheduler==3.7.0
aiohttp
aiofiles
bs4
emoji==1.6.0
gitpython
httpx[http2]
humanize
gtts
schedule
sqlalchemy==1.3.23
lottie
numpy
httpx
heroku3
requests==2.28.2
motor==3.0.0
opencv-python-headless
psutil
dnspython
psycopg2-binary
ptable
pyrogram
pymongo
pyromod==1.1.0
pytube
python-dotenv
pymediainfo
SpeechRecognition
prettytable
git+https://github.com/Itz-fork/py-trans.git
reportlab
svglib
speedtest-cli==2.1.3
tgcrypto 
textblob
tracemoepy
urllib3
uvloop
telegraph
uwuify
wget
pyaes==1.6.1
pysocks==1.7.1
aiosqlite>=0.16.0,<0.18.0
Pillow
git+https://github.com/Itz-fork/py-extract.git
ubotlibs
